refactor(sdk): move xml and metadata helpers into utils package
